Organise files by namespace
- Removed dead code.
