+++ src/main/game/Moves/Move.java
@@ -1,4 +1,4 @@
-package main.game;
+package main.game.Moves;
 
 public interface Move {
     Moves scoreAgainst(Moves playerTwoMove);
